Master_DataScience/M2_MachineLearning_Deploiement/serving/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
scikit-learn==1.0.2
imbalanced-learn == 0.11.0
numpy
Pillow
pandas==1.4.1
